requirements.txt
Flask>=3.0,<4
python-dotenv>=1.0
numpy>=1.24
plotly>=5.22
twilio>=9.0
gunicorn>=21.0
